refactor: move TMDB server client to services/ (no behavior change)
